Bug #37366: fix translating internal messages
svn r60749 introduced that the u.m.c.module was imported before the
locale was set. This caused that all Translation._() instances were bound to the unset locale.
The best way to fix this was moving that code into a subpackage. Also as
the Base class is now used for internal commands the naming is more
acurate to the real life situation.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-2@60871 1b283449-9f2f-0410-b571-8eb228e1a901