Bug #38823 test: Wait for pending transactions
When the context UCSTestUDM() is left, the UDM object is removed. If the
listener is too slow, it might not yet have processed the transactions.

The same happens if the context Listener() is left, which might
terminate the Listener before all pending transactions have been
processed.

Clone /usr/share/univention-directory-listener/get_notifier_id.py and
wait for transactions to be processed before checking for validity.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-3@64161 1b283449-9f2f-0410-b571-8eb228e1a901