Bug #39610 Up: Fix component server parsing
If only one of 'repository/{mirror,online{,/component/*}}/{srver,port}'
is set without the corresponding host/port entry, the component will use
the given component-server with the wrong repository-port, which fails.

Use the schema specific port, when only the hostname is given.
Still allow only port to be changed.

Make copy of base-url to not mess up global config.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-0@64865 1b283449-9f2f-0410-b571-8eb228e1a901