Bug #39957 listener: Disable DB sync while initializing module YAML
If a module fails to initialize or is installed later on, each entry in
the cache is updated to remove the module from the list of modules
registered for that entry. The cache is flushed after each change, which
adds a considerable delay to the initialization.

Disable the flushing while modules are initialized and flush the cache
once after all modules have finished.

YAML for r65674

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-0@65675 1b283449-9f2f-0410-b571-8eb228e1a901