Bug #35153 test: Kill process group on timeout
subprocess.Popen().terminate() only sends SIGKILL to the specific
process. If the process has forked children, they keep running. This is
quiet common for shell scripts, which fork lots of processes.
This leads to ucs-test not terminating, because it keep waiting for the
STDOUT, STDERR pipe to close, which they don't, because the children are
still considered valid writers.

Put the test into its own process group by making it a new session
leader, so that killing it will kill all it sub-processes as well.

Change _run_tee() to shutdown after 1s after the direct child has been
killed.

Move killing the child from signal handler to main-loop in _run_tee(),
as it calls sleep() during which the pipes are not read, which blocks
the child and prevents it from cleaning up properly before forcefully
being killed by the final SIGKILL.

The following test will fail with the old implementation, but will
terminate correctly with the new one:
    #!/usr/share/ucs-test/runner bash
    ## desc: Timeout 1
    ## exposure: safe
    ## timeout: 1
    trap "echo SIGTERM" TERM
    (
        trap "echo SIGTERM" TERM
        for ((i0=;i<10;i++));do sleep 1;done
    ) &
    ps f -o pid,ppid,pgid,sess,cmd
    sleep 3
    wait
    :

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-0@66268 1b283449-9f2f-0410-b571-8eb228e1a901