Bug #38051 LDAP: Initialize LDAP only on master
For slave LDAP servers the schema must be first replicated from the
master; otherwise they have an empty file
/var/lib/univention-ldap/schema.conf with not even core.schema included,
which makes "cn,c,l,o,ou" unknown. slapd then refuses to start because
it doesn't know the attributes used in "suffix" and "index".

This is normally fixed as soon as 03univention-directory-listener.inst
starts the Listener, which replicates the schema for the first time,
after which slapd starts fine.

But this still fails unknown attributes used for indexing.

Initializing LDAP only works on the master.

Fail if calling slapadd fails.

Ignore errors killing a not running slapd

Create the included files already in the postinst and only on slaves.

Fix spelling mistake.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-0@66712 1b283449-9f2f-0410-b571-8eb228e1a901