Bug #40338 up: Re-execute after release upgrade
This fixes the case were during the update the Updater is changed or
UCRV are changed in an incompatible manner.

This also allows us to remove the duplicate code as the next instance
will check for pending updates itself and sets the UCRV accordingly.

Move the code for checking if the UCRV 'update/available' needs to be
updated to the now common function.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-1@67587 1b283449-9f2f-0410-b571-8eb228e1a901