Bug #40600 UDL: Stop writing listener transaction log
/var/lib/univention-ldap/listener/listener is written by the "LDAP
server translog overlay module" or by the Listener for cascading.
It is read and truncated by the Notifier.
The Notifier is installed by default on the DC Master and on DC Backups.
On all other UCS system roles the file grown without an upper size
limit.

Only enable writing the file when UDN is installed.

Remove code to create an empty file normally created by UDN.

Remove unrelated comment.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-2@69904 1b283449-9f2f-0410-b571-8eb228e1a901