Bug #41261 UDL: Write notifier ID atomically
There is a small time windows when the Listener is terminated, where the
file "/var/lib/univention-directory-listener/notifier_id" is empty.
After a restart the UDL then jumps to the current ID, skipping all
intermediate transactions.

Write the current notifier ID to a new temporary file and rename it for
atomicy.

Minimap backport of r69905 for UCS-3.2-8

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-3.2-8@69919 1b283449-9f2f-0410-b571-8eb228e1a901