Bug #41261 UDL: Restore schema/id writing
Revert part of r69907, which removed
"/var/lib/univention-ldap/schema/id/id". It is still required by the
Notifier on DC Backups when chaining the notifications.

As the Notifier still uses the deprecated "dnotify" which triggers only
on DN_MODIFY (but not DM_RENAME), we need to modify the existing file.
As luckily the ID is monotonic and thus its string length is monotonic
as well, we can just overwrite the exiting file.
But the stdio C library's stream interface does not support opening an
existing file without overwriting, which also works when the files does
not yet exist. As such use the lower-level FD interface.

Drop useless 'key' parameter for cache_[sg]et_schema_id()

Add warning message is writing files fails.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-2@69980 1b283449-9f2f-0410-b571-8eb228e1a901