Bug #41261 UDL: Fix notifier ID not incrementing YAML
Fix regression introduced by r69980: The notifier ID needs to be
initialized once before the loop is started. Later on the ID gets
updates from the ID returned by the notifier.c:153:
> »···»···id = trans.cur.notify.id;
Otherwise the ID does not increment in case of moves, as the same ID is
queried again and again.

YAML for r69998

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-2@69999 1b283449-9f2f-0410-b571-8eb228e1a901