Bug #41580: evaluate the set of options from object classes
The set options of an instance are now "parsed" upon instanciation.
Basically they are set from the object classes of the object.

In save() self.old_options must not set to the default options if the object doesn't exists.
Therefore we also need to set self._exists = True after creating the object. And resetting it to False when removing the object.
This was wrong previously and makes it (a little bit more) possible to further work with objects after some operations.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-2@70446 1b283449-9f2f-0410-b571-8eb228e1a901