Bug #41715: use +0 syntax
Use +0 instead of read permissions. Read permissions are granted at the
very bottom by the rule "access to * by users read". This change
prevents the theoretical case that users with anonymous-bind can read
entries when anonymous bind is disabled if there is a rule like "access to * by * +scxd stop".

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-2@70851 1b283449-9f2f-0410-b571-8eb228e1a901