Bug #41180: remove write access to pseudo-attribute 'entry' of cn=admin-settings,cn=univention
Write access to the pseudo attribute 'entry' allows the creation and
removal (and modrdn) of the object (if one has also write access to the
parents 'children' pseudo attribute). This is not needed, especially not
for regular users. If we would have had write access to 'children' of
cn=univention this would be a root exploit as one could remove the
entry and readd a posixAccount with uidNumber=0 instead.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-2@70863 1b283449-9f2f-0410-b571-8eb228e1a901