Bug #40458 up: Fix confusing update message about next blocked release
univention.updater.tools.UniventionUpdater.release_update_available()
will only return the immediat next available UCS release respecting
blocking packages. But the UMC code expects it to return the latest
available UCS release and uses that information for some very confusing
message.

Remove the UCS release version from those strings and reformat for
clarity.

This needs-backporting to 4.0, 3.3 and 3.2, but this has to be done
carefully because of Bug #40338.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-2@71270 1b283449-9f2f-0410-b571-8eb228e1a901