Bug #42029 Listener: Fix deletion of ACLed entries
In the rename-case the old cache is copied in line 725, but
change_delete_dn() does a 2nd cache lookup on its own based on the new
location, where (most likely) no entry is found.

cache_get_entry_lower_upper() then returns an empty cache entry, for which no
handlers are registered.

handlers_delete() then only calls the replication module, as it is hard-coded
and will print the "skipped" message for all other handlers.

As we already have fetched the cache entry, change change_delete_dn() to
work directly on the fetched cache entry. As the cache entry is now
passed directly, change the name to change_delete() and make it static,
as it's only used within one compilation unit.

Merge r71776 from UCS-4.1-3

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@71779 1b283449-9f2f-0410-b571-8eb228e1a901