Bug #42586: ucs-test (appcenter): rework 79_umc_broken_failure_error_exit_code
Instead of expecting the actual installation/uninstallation to fail, this test
now tests if the `appcenter/progress` call returns the dpkg errors. In addition
it tests whether the app is re-installable.

The tests capture the state before the APP-Center-Switch. After the switch the
places marked with `FIXME (APPCENTER-SWITCH)` should be adapted.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@72981 1b283449-9f2f-0410-b571-8eb228e1a901