Bug #41735 ucr: Check checking non-existing files
All packages using "univention-install-config-registry" move their UCR
files to "/etc/univention/templates/removed/" in their .prerm and
restore them in .preinst and .postinst scripts. So while a package is
removed (but not purged), its config file is not there, but still liste
in the dpkg/status file. As such is is checked by
'univention-check-template' and considered modified.

Skip not existing files. If the package is ever re-installed, a modified
template file will be detected then.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-3@73311 1b283449-9f2f-0410-b571-8eb228e1a901