Bug #39068 join: Use hostname from LDAP
If the computer object for an UCS host is pre-created before the host
ist joined, the casing might be different. This might break downloading
the certificate for "$hostname.$domainname", as UNIX file systems are
case-ware traditionally.

Overwrite the local hostname with the name returned by the LDAP query to
either create a new entry or modify an already existing host.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-4.1-4@74974 1b283449-9f2f-0410-b571-8eb228e1a901