Bug #43217 dns: Fix registration of UCS-domain and forward servers
A DNS server must either be a UCS-domain name server, or an external
(resolving) domain-name server. In an external DNS server is used as a
UCS-domain name server, it can't answer UCS-domain related queries and
will return a negative answer, which breaks random services.

Remove broken shell code from 90univention-bind-post.inst and
re-implement is in Python:
- it didn't remove the forwarders from the list of name-servers.
- it only matched UCRV 'nameserverX' to 'dns/forwarderX', but not the
  other ones.
- it used the local "default IP" instead of "localhost", which does not
  work when the network is re-configured.
- it never heard of IPv6.
- it assigned "nameserver4", which is used nowhere.
- it was not idempotent.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@75433 1b283449-9f2f-0410-b571-8eb228e1a901