Bug #42380 runit: clamp Max-delay to 5s
During the upgrade to UCS-4.2 services like univention-bind may fail to
restart because of required pending upgrades (Bug #43200). In that case
runit tries to restart that service but fails each time, incrementing
the delay between restarts up to 60s. If then the service is restarted
through "service $SERVICE restart", the sleep process is NOT terminated
and "sv restart" aborts after 7 seconds - the restarting script then
continues and often fails as the service is not yet ready.

Long-term we want to replace runit by systemd anyway, so for now reduce
MAX_DELAY to 5 seconds.

And use "exec" for the sleep so that "sv status" shows the PID of the
sleep process and not of the intermediate shell.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@75516 1b283449-9f2f-0410-b571-8eb228e1a901