Bug #43217 DNS: Add --force-self
When univention-bind is installed and later joined, use the local
server as the preferred DNS server.

Only so this after joining into the domain, as only then the local LDAP
is populated and contains the same zones as on the master.

When dns/external is enabled, univention-fix-ucr-dns will quit without
doing anything.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@77374 1b283449-9f2f-0410-b571-8eb228e1a901