Bug #43919 krb: Remove pidfile annotation
Due the that "My samba4 KDC is more important than your Heimdal KDC"
renaming of the univention-keimdal.univention-service file, the generic
UCRV */autostart handling does not work for Heimdal. Therefore
heimdal-kdc.service is not masked and when started, the init script
aborts with "exit 0" due to "kerberos/autostart=no". But as no PID file
is created, systemd considers the service as failed.

The init script also starts multiple service, so using only one PID file
is "wrong" anyway.

So for now remove the annotation again until is is converted to a
systemd service anyway.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@77956 1b283449-9f2f-0410-b571-8eb228e1a901