* Bug #43594: fix concatenation and merge issues
PO can now be concatenated, which enables us to build a single PO
template from mixed source types, e.g. HTML and JS files in
univention-management-console.
The join_existing function is an alternative to achieve this, which
doesn't modify the header the way msgcat does it.
In the future it would be beneficial to move those catalog operations to
dh_umc completely and not rely on message_catalogs at all.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@77964 1b283449-9f2f-0410-b571-8eb228e1a901