Bug #44037 test/up: Clean "update in process APT source"
Bug #44346 introduced a work-around to copy the "temporary update
sources list" to "update in process list", as otherwise after a failed
update the new packages are no longer available.

This introduces a regression in ucs-test for the updater, as test 54
contains such a test case, where the update fails. This leaves behind a
temporary source entry, which is no longer valid afterwards.

Remove any "update in process list" after each test.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-0@80019 1b283449-9f2f-0410-b571-8eb228e1a901