Bug #45101 NFS: Start NFS server for first NFS share
The SysV-init-script "exit 0"s when no NFS share is exported. This leads
to systemd considering the 'nfs-kernel-server.service' to be running,
while not 'nfsd' was started. Any later 'exportfs -r' will do nothing
and a 'start' of the service is ignored, because for 'systemd' the
service is already running.

By setting 'RemainAfterExit=no' we tell systemd to consider the service
unit as 'inactive' instead when no 'Type=forking' process remains after
starting. 'GuessMainPID=yes' will detect 'rcp.mountd' as the main
process. This is not 100% correct, as the 'nfsd' kernel threads are the
real threads, but that should be okay.

nfs-utils from Debian-Stretch is fixed and provides native systemd unit
files.

git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.2/ucs-4.2-1@81554 1b283449-9f2f-0410-b571-8eb228e1a901