Bug #43745: remove existing _domaincontroller_master._tcp SRV record before adding another entry
If the DC Master should be replaced with another DC Master the old
record is still in the DNS of the Active Directory server. Only the old
DC Master has permissions to modify/remove it. So we are doing a cleanup
with Domain Administrator credentials and add a new SRV record as the
machine account.