Bug #45842: move self.modifypassword logic into the place where it is used
self.oldattr can never have a password entry (as these are LDAP
attributes, the real attribute is named userPassword).

The reason why this code is so hacky must be to make it possible to
lock a user which has no defined userPassword.

Our current code therefore sets a password value = '********' if the
user has no password defined or if the user which modifies this user has
no permissions to read the userPassword.