Bug #45842: Fix a re-regression in users/user open if self.exists()
amending 13.0.19-23 by considering the refactoring steps of 13.0.2-1
in more detail. If self.exists in open() then we need to save()
after _load_groups(). If not, then not (that was 13.0.19-22).

also add a test to ucs-test checking that membership attributes get
removed from additional groups the group membership is removed from the
user.