Bug #44621: Fix LDAP INVALID_CREDENTIALS with SAML
Throw 401 in case of invald ldap credentials. The saml message
is used as a password. But that message is only valid for 5 min.
A 401 does a passive saml login and sets the new message as the
new password. The ldap connection cache needs to be invalidated
because it is not possible to change the bind password which is
used for auto reconnect (ReconnectLDAPObject).

Impoved passive saml login in frontend.