Bug #46993 test: Drop OpenDNS resolver
Use only the AmazonProvidedDNS.

Due to Bug #44462 we have to save /etc/resolv.conf before unsetting the
UCR variables as the DHCP provided DNS servers are NOT copied to UCR; as
soon as we touch nameserverX /etc/resolv.conf will be re-generated by
UCR and we will loose even the DHCP provided servers.

To work around this we copy the remaining (AmazonProvidedDNS) to
nameserver1 and dns/forwarder1 as well to also prevent Bug #39807.