Bug #41772: Do not create empty snapshots
The old code requires at least one qcow2 image for taking a live
snapshot, as the 1st qcow2 file is used to store the runtime state.

But taking a snapshot of a VM with no HD does not work either, as
libvirt prevents this. Theoretically this could just snapshot the XML
data, but practically is does not work.