Bug #47260: Do not ignore ucr layers in ucr autostart module
The ucr autostart module used the changed variable which it got from ucr
to decide if a service needs to be changed. The changed variable does
not hold the effective change but the change on a layer.

The autostart module now always changes the service according to the
effective ucr value.