Bug #42022: Apply patch from QA
What the QA said about the patch:
* The errors reported in the CHECK_RUN phase of univention-join are not
written to join.log but only to stdout. The attached patch would fix
both: a) log to join.log too and b) display the error message. If you
are installing from a DVD it's pretty hard for the user to retrieve the
join.log from that system (e.g. no ssh yet).

* The attached patch additionally adds another binddn search via GSSAPI
that could help avoid the issue reported on this bug itself: Currently,
in case the udm search fails, we only fall back to ldapsearch via LDAPI,
which only works for root, and finally anonymous LDAP search, which is
disabled by default. Both isn't likely to succeed. The GSSAPI search may
also fail, but it's at least something that could possibly work for
users != root.