Bug #41109: do not always update univentionLastUsedValue for uidNumber
Normally, the value univentionLastUsedValue for uidNumber and gidNumber
increases automatically and monotonously when creating new users/groups.
The value was saved in LDAP to optimize the search for a free UID/GID number.

However, a smaller value was saved in LDAP if a smaller value for
uidNumber/gidNumber was explicitly specified when creating a user/group
(in short: the last used value was always saved).
This could mean that the next used UID/GID numbers were smaller than the
largest UID/GID number assigned so far. The numbers were reused if they
were free. This caused problems in some environments.

Therefore, this commit changes the behavior so that the last used value
is only stored in LDAP if it was determined automatically and NOT specified
manually.