Bug #47264: Add section on necessary default changes
Removed the hint on the entry point. The provider portal says:
"CMD for the Docker App. If Entrypoint is defined in the Docker Image,
these are the parameters for it when starting the App. If not, this is
the complete command that is run when the App is started. Leave emtpy
for image defaults."

The default therefore is to leave it empty. No change needed for the
developer in the portal.