Bug #47700: fix saml metadata creation for umc
In case a certificate chain is set to be used with the umc, the wrong
certificate was extracted during metadata creation for saml. The
de facto standard is to start the chain with the server certificate. For
the metadata creation the last one was extracted instead, which resulted
in signature validation errors on the IdP site.