Bug #47956: Fix disk target dev name assignment
Qemu does HW virtualization and this provides
- multiple IDE/SCSI/FD/SATA controllers
- with multiple buses
- with multiple targets
- with multiple units
How those devices are named in the guest OS cannot be controlled by
Qemu. As such the name given as disk/target/@dev is only used by libvirt
to order the devices by name and then assign those 4-tuple addresses per
bus. If @dev is later changes, this does not re-assign addresses.

This results in an duplicate address being assigned when for an existing
device @dev is changes, address is kept and a new device is added
re-using the old @dev name.

Re-implement the logic from libvirt/conf/domain_conf.c in UVMM to
calculate the address in advance and skip it if it is already assigned.

Note: The alternative to overwrite @dev when address is specified does
not work as at least both SCSI and SATA use the prefix "sd" for naming
their devices, so the mapping is not unique.

Move the code of assign_disks() from umc/python/uvmm/domains.py to
src/univention/uvmm/storage.py and remove Disk.address again.