Bug #48349 kernel: Always Update UCRV update/reboot/required"
As the hook script is a conffile in /etc/kernel/postrm.d/, it is only
active _after_ univention-kernel-image has been configured. As the
package pulls in "linux-image-$ABI" as a dependency, the kernel is
already configured and the hook is not yet called.

Restore the package postint maintainer script to set the variable there
too when:
- it is not a container
- unattended-upgrade created the flag file "/var/run/reboot-required"
- the running kernel is not the installed kernel
- this is a new installation or an upgrade, which pulled in a new
  kernel.

Future kernel updates will use the hook script IF this u-k-i update is
already installed.
If u-k-i is only installed TOGETHER with a new FUTURE kernel, the ABI
code will detect this.