Bug #48427 ldap: Add transaction log database
Split the slapd module loading chunk into a separate file to allow
loading other modules.

Put the code to setup the cn=translog database into a separate file to
be able to call it on
1) upgrades of joined master and backup
2) new installation of master
3) joining backups
without running the join-script "01" again.

Hide slapcat output.
Force remove translog on re-join.

For performance reasons we will sacrifice 'D' from ACID and will flush
the database from memory to disk only every minute.
When next UDN is started make sure the last transaction was imported.

Also check that the cn=translog is actually initialized as "slapd"
creates an empty database on startup if the files are missing - but the
base entry will be missing and any subsequent add will fail.