Bug #48427 join: Copy translog.mdb on Backup join
so UDN on that Backup knows the same transactions as the Master.
Remove old index file - it may be invalid.
Use importer in no-index mode.
Fix spelling in two debug messages.

Hide slapcat output.
Force remove translog on re-join.

If the database is corrupt, it must be purged before "slapd" can be
restarted again, which happens
1) in univention-join:854 itself
2) in 01univention-ldap-server-init.inst:116
3) in 03univention-directory-listener.inst:205
4) indirectly by replication.py
before that post-hunk-code was executed.

But univention-join is a mess:
`slapd` is started before the LDAP SCHEMA got replicated on systems
other than the master. There is code to fake the initial schema in
`01univention-ldap-server-init.inst`, but this is too late.

For now we add back the code to initialize the `cn=translog` database
late and ignore the error of the purge.