"Bug #48427 join: Move code to force-reset translog DB bask
univention-join is a mess:
`slapd` is started before the LDAP SCHEMA got replicated on systems
other than the master. There is code to fake the initial schema in
`01univention-ldap-server-init.inst`, but this is too late.

For now we add back the code to initialize the `cn=translog` database
late and ignore the error of the purge.