Bug #48427: udn,ldap: Performance issues
For performance reasons we will sacrifice 'D' from ACID and will flush
the database from memory to disk only every minute.
When next UDN is started make sure the last transaction was imported.

Also check that the cn=translog is actually initialized as "slapd"
creates an empty database on startup if the files are missing - but the
base entry will be missing and any subsequent add will fail.