Bug #48910 up: Improve iterating over releases
When `univention-repository-create` is called for UCS-4.3-3, the old
code started with 4.0-0 and tried to reach 4.3-3 via 4.1-0, 4.2-0 and
4.3-0. As the last does not exists when the initial repository is
created with the UCS-4.3-3 DVD, 4.3-3 was never reached.

Split _iterate_versions() into two functions:
- _iterate_releases() just yield the UCS releases from start to end
including. It uses a generator() where the caller can provide feedback
using the send(failed): If the (patch-level-)release is not found, the
iterator then can skip to the next minor or major version.
It also makes sure to always iterate all version for the end release,
that is for 4.4-1 it will try 4.0-0, 4.1-0, 4.2-0, 4.3-0, 4.4-0 and
4.4-1.

- _iterate_releases() uses that function and does the checking using
access().

A proper unnit test for _iterate_releases() is added.