Bug #48126 USS: Fix proxy checking
For simplicity I would use "curl -x" to specify the proxy via the
command line, but as the UCRV may contain a
- http:// URL
- https:// URL
- a plain host name
different proxy settings must be used. On top there is "no_proxy".
So instead export all "proxy" settings via the process environment and
let "curl" do the right ting.

Also add "-I" to do an HEAD request instead of a GET request.

Also log the output of the failed command to help debugging failures.