Bug #28233 udn: Fix index types
Use *un*signed `size_t` and use correct prefix in printf() to dump it.

The correct type would be "off_t" for file-offsets, but that breaks the
ABI for i386 as the size depends on _FILE_OFFSET_BITS: Currently we
compile UDN with _FILE_OFFSET_BITS=64, but as we used size_t which is
only 32 bit on i386, the change would break on i386.