Bug #49520 ucslint: Fix ucslint.override application
The "./debian/ucslint.overrides" file contains path pattern, which are
interpreted relative to the package directory.

But "ucslint" can be invoked with a path - by default it defaults to the
current working directory - in which case tests walking the file system
will use the path relative to the current working directory, which then
will not match as they have an additional path prefix.

Change the current working directory to the given path before running
the test and always use "." as the current directory.

The paths to the plugin-directories must be resolved before chdir() is
used.

Later on use "normdir()" instead of "absdir()" to just normalize the
paths for comparison.

Output line number when parsing the file "debian/ucslint.overrides".