Bug #49704 ucslint: Move to univention namespace
If "ucslint" is installed while building a new version of ucslint, the
installed (old) version is used, which may fail.

Move all files to the univention/ sub-directory and use that as
otherwise calling bin/ucslint from the checked out source tree will not
run the UCS specific plugins, as they were in ucslint-univention/ which
is not included by default.