Bug #49695: Handle error during UVMM schema registration
The old package version provided the schema and ACL as files on the
master as UCR template fragments. While the template and info files are
removed, they remain in /etc/ldap/slapd.conf until the file is
committed. This may leads to slapd to being able to restart as the ACL
then referenced the not existing attributes 'univentionVirtualMachine*'.

Make sure to re-generate the file and only execute the code once on
upgrade only.

Log error message if slapd and/or UDL are not running during the upgrade
as then registering the schema/ACL may fail.