OpenLDAP: Change certificate location to /etc/ldap/ssl
/etc/univention/ssl is only permitted to the user root.

By executing slapd as user openldap it's easier to do the same as
for FreeRADIUS (for example).

This doesn't fix yet the switchover from running under root as
currently to openldap.

At this point, when the certificates are copied into the right place
and given the right permissions, slapd starts working again.