Bug #47014 EC2 image: Truncate machine-id
DBUS introduced the concept of a machine ID, which never changes and is
unique per instance: each clone must have a new ID.
Later systemd generalized this, but the same applies there, too.

The ID is used by networking services, desktop environments, the journal
and others.

The correct way to generate a new one is to truncate the file site to 0.

See <man:machine-id(5)> and <man:systemd-machine-id-setup(1)> and
<https://wiki.debian.org/MachineId>