Bug #49872 udn: Fix `univention-translog stat` check for index
Since Bug #48729 we allow transactions to be purged from the log; it no
longer starts with TID 1. So the *count* of transactions in the log is now
less then the *last* transaction number.

On the other hand the index still contains *all* transactions. So the
index is okay when it matches the *last* transaction, not the *count*.