Bug #27816: Bug #50025: the check for a valid superordinate is now delayed to the creation or modification of objects.
It's otherwise not possible to create a instance of a object() which
requires a superordinate object.
This delays the requirement to the point when it's acutally really
needed - before create() or modify().

The validation was added too strictly in git:5612aecb7e6439a51c8888a0a7f6dafc0305585f
via Bug #34764 ;-)