Bug #49967 nagios: Set PATH in SUID wrappers
exec?e(envp=NULL) only sets `PATH=/usr/bin:/bin` (or whatever
`confstr(_CS_PATH)` returns. This is not enough as we need to execute
programs from `/sbin:/usr/sbin` as well.

As it is the job of the SUID wrapper to provide a clean and sane
environment, explicitly configure PATH.

Also fix debian/control to fix several warning:
    dpkg-gencontrol: Warnung: Feld Provides von Paket univention-nagios-server: unbekannte Substitutionsvariable ${python:Provides}
    dpkg-gencontrol: Warnung: Paket univention-nagios-client: unbenutzte Substitutionsvariable ${shlibs:Depends}
    dpkg-gencontrol: Warnung: Paket univention-nagios-client: unbenutzte Substitutionsvariable ${shlibs:Depends}
    dpkg-gencontrol: Warnung: Feld Depends von Paket univention-nagios-common: unbekannte Substitutionsvariable ${python:Depends}

Turn on build hardening.

Fixes: 62d22bd0a4eb08ca3ebb8aabf0e139c08e6824d4