Bug #50213: Rebuild signed GRUB 2.02~beta3-5+deb9u2A~4.3.5.201910070951
The source package `grub-efi-amd64-signed` started as a native package,
but was later converted to a UCS package, but did NOT use the 3 parts
schema. Because of that the version was NOT bumped for UCS-4.4, so the
2.0.0-3 is already taken by UCS-4.4-2, which now blocks the import for
UCS-4.3-5.

'+' sorts too early, so just skip '-3' and continue with '-4' as the
GRUB2 version number is inserted anyway.

Also fix the regular expressions to strip the suffix used since UCS-4.