Bug #42080: don't modify primary group for users/user during open()
* the primary group might not be replicated yet
* the user might have no permission to read his primary group

If we don't find a primary group for the gidNumber opening a UDM module should just be possible, while `primaryGroup` is unset then.
As `primaryGroup` is required we can set it to the default value before modify() so that the old behavior is delayed until a modification:
it still raises `permissionDenied` and prevents modification of users if permission or replication problems exists.
`permissionDenied` is of course then not raised for administrator accounts with write access to e.g. the default primary group "Domain Users" and in that case the primary group would still be changed/reset.
This way the change is at least not so invasive as an complete API change would be.

Fixes: Bug #12621
Fixes: 5d3da2a0e701a082ef12072e186cd5ce5fce696d
Fixes: 372ce7317e507ed05dcebce0c7736746428ed7b5