Bug #50629: only apply diff via ADD or DELETE when modifying group memberships
remove all members with ldap.MOD_DELETE and add all new members with ldap.MOD_ADD.

The code was prone to race conditions if there are changes inbetween on AD side: causing a loss of group members.
Especially in large environments with a lot of group members this might have an performance impact.