AD connector: Make UPN configurable
By introducing a new UCR variable instead of mapping uid@KERBEROSREALM from
UCS OpenLDAP to Active Directory's userPrincipalName we can now map a specific
attribute from UCS OpenLDAP as the resulting userPrincipalName in Active Directory.

However it is up to the administrator to use an attribute that is both single-valued
and complies with RFC 822 as that is what Microsoft expects for userPricipalName.
Hence this option is not defined by default and the defaults remain sane.